chore: move the observable procedure model out of its own directory
