+++ core/observable_procedure_map.ts
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {IProcedureMap} from '../interfaces/i_procedure_map.js';
-import type {IProcedureModel} from '../interfaces/i_procedure_model.js';
-import {isObservable} from '../interfaces/i_observable.js';
+import {IProcedureMap} from './interfaces/i_procedure_map.js';
+import type {IProcedureModel} from './interfaces/i_procedure_model.js';
+import {isObservable} from './interfaces/i_observable.js';
 
 
 export class ObservableProcedureMap extends
