+++ core/procedures.ts
@@ -29,7 +29,7 @@ import {IParameterModel} from './interfaces/i_parameter_model.js';
 import {IProcedureMap} from './interfaces/i_procedure_map.js';
 import {IProcedureModel} from './interfaces/i_procedure_model.js';
 import {IProcedureBlock, isProcedureBlock} from './interfaces/i_procedure_block.js';
-import {ObservableProcedureMap} from './procedures/observable_procedure_map.js';
+import {ObservableProcedureMap} from './observable_procedure_map.js';
 import * as utilsXml from './utils/xml.js';
 import * as Variables from './variables.js';
 import type {Workspace} from './workspace.js';
